requirements.txt
Authlib==0.14.1
boto3==1.12.26
botocore==1.15.26
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
configobj==5.0.6
cryptography==2.8
docutils==0.15.2
fsspec==0.6.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
gspread==3.3.0
gspread-dataframe==3.0.4
httplib2==0.17.0
idna==2.9
jmespath==0.9.5
numpy==1.18.2
pandas==1.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
s3fs==0.4.0
s3transfer==0.3.3
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
